:truck: [#29] move drf-spectacular extensions into separate files
