+++ vng_api_common/extensions/polymorphic.py
@@ -1,22 +1,8 @@
 from drf_spectacular.extensions import OpenApiSerializerExtension
-from drf_spectacular.openapi import AutoSchema
 from drf_spectacular.plumbing import ResolvedComponent
+from drf_spectacular.settings import spectacular_settings
 
-from vng_api_common.utils import underscore_to_camel
-
-
-class GegevensGroepFieldExtension(OpenApiSerializerExtension):
-    target_class = "vng_api_common.serializers.GegevensGroepSerializer"
-    match_subclasses = True
-
-    def map_serializer(self, auto_schema: AutoSchema, direction):
-        schema = auto_schema._map_serializer(
-            self.target, direction, bypass_extensions=True
-        )
-
-        del schema["description"]
-
-        return schema
+from ..utils import underscore_to_camel
 
 
 class PolymorphicSerializerExtension(OpenApiSerializerExtension):
@@ -36,6 +22,8 @@ class PolymorphicSerializerExtension(OpenApiSerializerExtension):
             self.target, direction, bypass_extensions=True
         )
         base_name = f"Base_{self.target.__class__.__name__}"
+        if direction == "request" and spectacular_settings.COMPONENT_SPLIT_REQUEST:
+            base_name = base_name + "Request"
         base_component = ResolvedComponent(
             name=base_name,
             type=ResolvedComponent.SCHEMA,
@@ -56,6 +44,8 @@ class PolymorphicSerializerExtension(OpenApiSerializerExtension):
                 schema = {"allOf": [base_component.ref, sub_component.ref]}
 
             component_name = f"{resource_type.value}_{self.target.__class__.__name__}"
+            if direction == "request" and spectacular_settings.COMPONENT_SPLIT_REQUEST:
+                component_name = component_name + "Request"
             component = ResolvedComponent(
                 name=component_name,
                 type=ResolvedComponent.SCHEMA,
